Moved method 'subclasses' to an external file
